+++ src/analytics/index.js
@@ -24,8 +24,8 @@ export {
   trackMasteryTierChanged,
   trackStepMismatchClicked,
   trackSupportLinkClicked,
+  trackAffiliateLinkClicked,
   trackLocaleChanged,
 } from "./track.js";
 export { installErrorCapture, sendOnPageHide } from "./transport.js";
-export { installWebAnalyticsPresence } from "./web-presence.js";
 export { installLocaleEngagementTracking } from "./locale-engagement.js";